Is this an idiom too?
I'm really not sure if this is an idiom. ...not an English native speaker.

This is the sentence at line 162: "Pattern observers are a simple and flexible that will allow you to observe your data any way that you want."

I think a substantive is missing  => "are a simple and flexible"...?

But, again, sorry if my assumptions are wrong.