Debug initializer
Maybe this is the wrong place? I couldn't find documentation for `DEBUG`, so I threw it in here. It *is* in the migration docs, but I think it needs to be linked somewhere in the actual docs, not just migration. If you think this needs to go somewhere else, I'd be happy to do a different PR.